+++ resources/js/wirekit-alpine.js
@@ -53,6 +53,7 @@ import wirekitTableSort from './components/table-sort.js';
 import wirekitTour from './components/tour.js';
 import wirekitResizableHandle from './components/resizable.js';
 import wirekitImageCompare from './components/image-compare.js';
+import wirekitLightbox from './components/lightbox.js';
 import wirekitStatAnimate from './components/stat-animate.js';
 import wirekitAnimate from './components/animate.js';
 import wirekitReadingSpine from './components/reading-spine.js';
@@ -112,6 +113,7 @@ if (alreadyHasAlpine()) {
     Alpine.data('wirekitTour', wirekitTour);
     Alpine.data('wirekitResizableHandle', wirekitResizableHandle);
     Alpine.data('wirekitImageCompare', wirekitImageCompare);
+    Alpine.data('wirekitLightbox', wirekitLightbox);
     Alpine.data('wirekitStatAnimate', wirekitStatAnimate);
     Alpine.data('wirekitAnimate', wirekitAnimate);
     Alpine.data('wirekitReadingSpine', wirekitReadingSpine);